refactor(member): 회원 약관 흐름 Facade 분리
회원 생성과 추가정보 입력에서 약관 동의 저장/검증 조합을 MemberFacade로 옮겨 MemberCommandService가 회원 상태 변경에 집중하도록 정리합니다.

Constraint: MemberCommandService가 MemberTermsFacade를 의존하는 방향은 모듈 내부 역할상 부적절

Rejected: MemberCommandService 안에서 약관 Command/Query를 직접 조합 | 회원 명령 서비스 책임이 커짐

Confidence: high

Scope-risk: moderate

Tested: ./gradlew test --tests checkmo.member.MemberApiTest --tests checkmo.member.TermsDomainServiceTest --tests checkmo.authentication.AuthSignupTermsApiTest --tests checkmo.authentication.AuthSignupTermsStrictApiTest --tests checkmo.CheckmoApplicationTests

Not-tested: 전체 ./gradlew test
